Fix issue where domain is provided but zone id is not.
There are some cases where you may want to create the custom domain and
base path mapping, but not the DNS record. Generally that is not the
case, but if you are using an outside process to attach a custom domain
to a private API, you will probably not want this record generated here.